[dev] refactor(ui): 重构应用布局和头部组件
将头部导航栏逻辑提取为独立组件 AppHeader
优化应用布局结构，修复滚动和高度问题
调整 CSS 样式以改善用户体验
