+++ prompto-lab-ui/src/assets/base.css
@@ -59,7 +59,7 @@
 }
 
 body {
-  min-height: 100vh;
+  height: 100vh;
   color: var(--color-text);
   background: var(--color-background);
   transition:
