+++ prompto-lab-ui/src/assets/main.css
@@ -17,7 +17,7 @@ body {
 }
 
 #app {
-  min-height: 100vh;
+  height: 100vh;
   width: 100%;
 }
